removed duplicate CheckInActivity.java file
