+++ app/src/main/AndroidManifest.xml
@@ -31,7 +31,7 @@
         </activity>
         <activity android:name=".CourseRegistrationActivity" />
         <activity android:name=".CourseDetailsActivity"/>
-        <activity android:name=".CheckInActivity"/>
+        <activity android:name=".ui.student.CheckInActivity"/>
         <activity android:name=".MyCoursesActivity"/>
         <activity android:name=".ui.instructor.HomeActivity"/>
